requirements.txt
Flask
Gunicorn
requests
urlextract
google.generativeai
os